修改如下： 1.增加了消息返回接口。 2.修改了Md5算法。 3.增加了停止终止定位接口。 4.MainApp方面进行了完善

+++ commonSrc/tools/systemenc/md5.cpp
@@ -83,22 +83,22 @@ documentation and/or software.
 Rotation is separate from addition to prevent recomputation.
 */
 #define FF(a, b, c, d, x, s, ac) { \
-	(a) += F((b), (c), (d)) + (x)+(unsigned long int)(ac); \
+	(a) += F((b), (c), (d)) + (x)+(uint32_t)(ac); \
 	(a) = ROTATE_LEFT((a), (s)); \
 	(a) += (b); \
 }
 #define GG(a, b, c, d, x, s, ac) { \
-	(a) += G((b), (c), (d)) + (x)+(unsigned long int)(ac); \
+	(a) += G((b), (c), (d)) + (x)+(uint32_t)(ac); \
 	(a) = ROTATE_LEFT((a), (s)); \
 	(a) += (b); \
 }
 #define HH(a, b, c, d, x, s, ac) { \
-	(a) += H((b), (c), (d)) + (x)+(unsigned long int)(ac); \
+	(a) += H((b), (c), (d)) + (x)+(uint32_t)(ac); \
 	(a) = ROTATE_LEFT((a), (s)); \
 	(a) += (b); \
 }
 #define II(a, b, c, d, x, s, ac) { \
-	(a) += I((b), (c), (d)) + (x)+(unsigned long int)(ac); \
+	(a) += I((b), (c), (d)) + (x)+(uint32_t)(ac); \
 	(a) = ROTATE_LEFT((a), (s)); \
 	(a) += (b); \
 }
@@ -232,10 +232,10 @@ void MD5_CTX::MD5Update(unsigned char *input, unsigned int inputLen)
 	index = (unsigned int)((this->count[0] >> 3) & 0x3F);
 
 	/* Update number of bits */
-	if ((this->count[0] += ((unsigned long int)inputLen << 3))
-		< ((unsigned long int)inputLen << 3))
+	if ((this->count[0] += ((uint32_t)inputLen << 3))
+		< ((uint32_t)inputLen << 3))
 		this->count[1]++;
-	this->count[1] += ((unsigned long int)inputLen >> 29);
+	this->count[1] += ((uint32_t)inputLen >> 29);
 
 	partLen = 64 - index;
 
@@ -288,9 +288,9 @@ void MD5_CTX::MD5Final(unsigned char digest[16])
 
 /* MD5 basic transformation. Transforms state based on block.
 */
-void MD5_CTX::MD5Transform(unsigned long int state[4], unsigned char block[64])
+void MD5_CTX::MD5Transform(uint32_t state[4], unsigned char block[64])
 {
-	unsigned long int a = state[0], b = state[1], c = state[2], d = state[3], x[16];
+	uint32_t a = state[0], b = state[1], c = state[2], d = state[3], x[16];
 
 	Decode(x, block, 64);
 
@@ -376,10 +376,10 @@ void MD5_CTX::MD5Transform(unsigned long int state[4], unsigned char block[64])
 	MD5_memset((unsigned char*)x, 0, sizeof (x));
 }
 
-/* Encodes input (unsigned long int) into output (unsigned char). Assumes len is
+/* Encodes input (uint32_t) into output (unsigned char). Assumes len is
 a multiple of 4.
 */
-void MD5_CTX::Encode(unsigned char *output, unsigned long int *input, unsigned int len)
+void MD5_CTX::Encode(unsigned char *output, uint32_t *input, unsigned int len)
 {
 	unsigned int i, j;
 
@@ -391,16 +391,16 @@ void MD5_CTX::Encode(unsigned char *output, unsigned long int *input, unsigned i
 	}
 }
 
-/* Decodes input (unsigned char) into output (unsigned long int). Assumes len is
+/* Decodes input (unsigned char) into output (uint32_t). Assumes len is
 a multiple of 4.
 */
-void MD5_CTX::Decode(unsigned long int *output, unsigned char *input, unsigned int len)
+void MD5_CTX::Decode(uint32_t *output, unsigned char *input, unsigned int len)
 {
 	unsigned int i, j;
 
 	for (i = 0, j = 0; j < len; i++, j += 4)
-		output[i] = ((unsigned long int)input[j]) | (((unsigned long int)input[j + 1]) << 8) |
-		(((unsigned long int)input[j + 2]) << 16) | (((unsigned long int)input[j + 3]) << 24);
+		output[i] = ((uint32_t)input[j]) | (((uint32_t)input[j + 1]) << 8) |
+		(((uint32_t)input[j + 2]) << 16) | (((uint32_t)input[j + 3]) << 24);
 }
 
 /* Note: Replace "for loop" with standard memcpy if possible.
